+++ src/main/java/cart/controller/CartItemApiController.java
@@ -1,4 +1,4 @@
-package cart.ui;
+package cart.controller;
 
 import cart.application.CartItemService;
 import cart.domain.member.Member;
