+++ src/main/java/cart/exception/ControllerExceptionHandler.java
@@ -1,7 +1,5 @@
-package cart.ui;
+package cart.exception;
 
-import cart.exception.AuthenticationException;
-import cart.exception.CartItemException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
